first version of release workflow
it should either trigger when a PR is complated, or manually triggered with the commit SHA for a PR.

the PR must be labeled with either major, minor or patch, which will then bump the version with the "instruction label", for the affected packages.

also adding templates for issues, updated copilor instructions and bumped uv version in the setup-uv-job action.